requirements.txt
Keras>=2.2.4
Keras-Applications>=1.0.6
Keras-Preprocessing>=1.0.5
numpy>=1.15.4
Pillow>=5.4.1
scipy>=1.2.0
scikit-learn>=0.20.2
tensorflow>=1.12.0